requirements_passport.txt
absl-py==0.12.0
apache-beam==2.30.0
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
avro-python3==1.9.2.1
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==21.6.0
craft-text-detector==0.3.5
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
dask==2021.3.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.1.1
Django==3.2.5
djangorestframework==3.12.4
dm-tree==0.1.6
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
fastavro==1.4.2
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.13.0
gin-config==0.4.0
glob2==0.7
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.5.0
importlib-resources==5.2.0
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jeepney==0.6.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lmdb==1.2.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mrz==0.6.2
natsort==7.1.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.2.0.34
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
portalocker==2.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
readme-renderer==29.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-image==0.14.2
scikit-learn==0.24.2
scipy==1.1.0
SecretStorage==3.3.1
Send2Trash==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.1.0
tensorflow-model-optimization==0.6.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
tf-models-official==2.5.0
tf-slim==1.1.0
threadpoolctl==2.1.0
toolz==0.11.1
torch==1.9.0
torchvision==0.2.1
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
twine==3.4.1
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.1
